requirementss.txt
Flask
pdf2image
pytesseract
Pillow
PyPDF2
